Fix entity classes and getDataset functioN
