+++ getDataset.py
@@ -8,68 +8,61 @@ Created on Fri Oct  5 19:03:50 2018
 
 # importing required modules
 from zipfile import ZipFile
-import entity as e
+from Entity.Point import Point
+from Entity.User import User
+from Entity.Detection import Detection
 import os
 import re
 
-# specifying the zip file name 
-file_name = "Geolife Trajectories 1.3.zip"
-userList = []
-# opening the zip file in READ mode 
-with ZipFile(file_name, 'r') as zipfile:
-    # printing all the contents of the zip file 
-    for elem in zipfile.infolist():
-        filename = elem.filename
-        # print("\t"+filename)
-        if not (elem.is_dir()):
-            # print("Directory")
-            # else:
-            filepath, filename = os.path.split(elem.filename)
-            filepath, trajectory = os.path.split(filepath)
-            filepath, user = os.path.split(filepath)
-            if filename.endswith(".plt") and not (filename.startswith("._")):
-                # print("\t"+user)
-                curUserInList = None
-                for itemUser in userList:
-                    if itemUser.get_identifier() == user:
-                        # esiste gia
-                        curUserInList = itemUser
-                if not curUserInList:
-                    curUserInList = e.User(user)
-                    userList.append(curUserInList)
 
-                curFile = zipfile.open(elem.filename)
-                # line 1.. 5 are useless in this dataset
-                for i in range(5):
-                    line = curFile.readline()
-                line = curFile.readline()
-                curFileName = re.sub('\.plt$', '', filename)
-                pointList = []
-                # count=0 #number of point
-                while line:
-                    line = curFile.readline()
-                    line = line.decode('UTF-8')
-                    if line:
-                        # count +=1
-                        latitudine, longitudine, zero, altitudine, date, dateS, timeS = line.split(",")
-                        pointList.append(e.Point(latitudine, longitudine, date))
-                
-                # print("\tcount: {}".format(count))
-                #print("\t\t"+curFileName+"\n")
+def get_dataset():
+    # specifying the zip file name
+    file_name = "Geolife Trajectories 1.3.zip"
+    userlist = []
+    # opening the zip file in READ mode
+    with ZipFile(file_name, 'r') as zipfile:
+        # printing all the contents of the zip file
+        for elem in zipfile.infolist():
+            if not (elem.is_dir()):
+                # print("Directory")
+                # else:
+                filepath, filename = os.path.split(elem.filename)
+                filepath, trajectory = os.path.split(filepath)
+                filepath, user = os.path.split(filepath)
+                if filename.endswith(".plt") and not (filename.startswith("._")):
+                    # print("\t"+user)
+                    curuserinlist = None
+                    for itemUser in userlist:
+                        if itemUser.get_identifier() == user:
+                            # esiste gia
+                            curuserinlist = itemUser
+                    if not curuserinlist:
+                        curuserinlist = User(user)
+                        userlist.append(curuserinlist)
 
-                curUserInList.add_detection(e.Detection(curFileName, pointList))
-                
-            # elif filename.endswith("labels.txt"):
-            # print("\t\tlables")
-            # else:
-            # print("not my file")
-zipfile.close()
-#
-"""
-for elem in userList:
-    print(elem.get_identifier())
-    for detection in elem.get_detectionlist():
-        print(detection.get_timestamp())
-        for point in detection.get_pointlist():
-            print(point.get_latitude()+","+point.get_longitude()+","+point.get_timestamp())
-"""
+                    curfile = zipfile.open(elem.filename)
+                    # line 1.. 5 are useless in this dataset
+                    for i in range(5):
+                        curfile.readline()
+                    line = curfile.readline()
+                    curfilename = re.sub('\.plt$', '', filename)
+                    pointlist = []
+                    # count=0 #number of point
+                    while line:
+                        line = curfile.readline()
+                        line = line.decode('UTF-8')
+                        if line:
+                            # count +=1
+                            latitudine, longitudine, zero, altitudine, date, dates, times = line.split(",")
+                            pointlist.append(Point(latitudine, longitudine, date))
+
+                    # print("\tcount: {}".format(count))
+                    # print("\t\t"+curFileName+"\n")
+
+                    curuserinlist.add_detection(Detection(curfilename, pointlist))
+
+                # elif filename.endswith("labels.txt"):
+                # print("\t\tlables")
+                # else:
+                # print("not my file")
+    zipfile.close()
